+++ client/store/auth.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 import history from "../history";
-import { fetchCart, goAddShoppingItem } from "./cart";
+import { fetchCart, mergeGuestCartwDBCart } from "./cart";
 const TOKEN = "token";
 const CART = "cart";
 import { emptyCart } from "./cart";
@@ -58,31 +58,7 @@ export const authenticate =
 
       const { id } = auth;
 
-      //check if there are items in a guest cart that need to be added to the backend cart for this user
-      const frontEndCart = JSON.parse(window.localStorage.getItem(CART));
-
-      //check if there are items in the backend cart that need to be merged with front end cart.
-      const { data: backEndCart } = await axios.get(`/api/cart/${id}`, {
-        headers: {
-          authorization: res.data.token,
-        },
-      });
-
-      if (frontEndCart) {
-        //go add the frontend cart to backEnd, then update the cart in redux store with the new information from the backend.
-        await frontEndCart.forEach(async (frontEndProduct) => {
-          //get qty of the item that is already in this user's back end cart
-          let [productInBackEndCart] = backEndCart.filter(
-            (backEndProduct) => backEndProduct.id === frontEndProduct.id
-          );
-          let backEndQty = 0;
-          if (productInBackEndCart) {
-            backEndQty = productInBackEndCart.qty;
-          }
-          let newQuantity = backEndQty + frontEndProduct.qty;
-          await dispatch(goAddShoppingItem(frontEndProduct, id, newQuantity));
-        });
-      }
+      dispatch(mergeGuestCartwDBCart(id));
 
       //get backend cart for this user (this is still needed here in case there was nothing in the front end cart)
       dispatch(fetchCart(id));
